+++ zebra-chain/src/chain_tip/tests.rs
@@ -1,3 +1,4 @@
 #![allow(clippy::unwrap_in_result)]
 
 mod prop;
+mod vectors;