+++ zebra-chain/src/transaction.rs
@@ -1705,6 +1705,29 @@ impl Transaction {
             .map(|shielded_data| &mut shielded_data.value_balance)
     }
 
+    /// Modify the `value_balance` field from the `ironwood::ShieldedData` in this transaction,
+    /// regardless of version.
+    ///
+    /// See `ironwood_value_balance` for details.
+    pub fn ironwood_value_balance_mut(&mut self) -> Option<&mut Amount<NegativeAllowed>> {
+        match self {
+            Transaction::V6 {
+                ironwood_shielded_data: Some(ironwood_shielded_data),
+                ..
+            } => Some(&mut ironwood_shielded_data.data_mut().value_balance),
+
+            Transaction::V1 { .. }
+            | Transaction::V2 { .. }
+            | Transaction::V3 { .. }
+            | Transaction::V4 { .. }
+            | Transaction::V5 { .. }
+            | Transaction::V6 {
+                ironwood_shielded_data: None,
+                ..
+            } => None,
+        }
+    }
+
     /// Modify the `value_balance` field from the `sapling::ShieldedData` in this transaction,
     /// regardless of version.
     ///